requirements.txt
streamlit>=1.37.0
pandas>=2.2.0
numpy>=1.26.0
snowflake-connector-python>=3.7.0
chromadb>=0.5.0
openai>=1.36.0
tiktoken>=0.7.0
tenacity>=8.2.3
python-dotenv>=1.0.1
plotly>=5.22.0
networkx>=3.3
neo4j>=5.19.0
diskcache>=5.6.3
pyyaml>=6.0.1
